# openapi/admin.openapi.yaml
openapi: 3.1.0
info:
  title: LaTeX Renderer Administrator API
  version: 0.3.0
  description: Administrative API used by the protected Web UI and administration CLI.
servers:
  - url: https://latex-render.n624.jp/admin/api/v1
security:
  - accessAndAdmin: []
paths:
  /session:
    get:
      summary: Inspect the verified Cloudflare Access identity and invitation linkage
      security:
        - cloudflareAccess: []
      responses:
        "200":
          description: Current Access identity and linkage state
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
  /session/claim-subject:
    post:
      summary: Link the verified Cloudflare Access subject to its invited administrator
      description: The subject and email are read exclusively from the verified Access JWT. Any request body is ignored.
      security:
        - cloudflareAccess: []
      parameters:
        - in: header
          name: X-CSRF-Token
          required: true
          schema: { type: string, const: "1" }
      responses:
        "200":
          description: Linked, or already linked to the same identity
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: "#/components/schemas/Session" }
                  - type: object
                    required: [claimed]
                    properties:
                      claimed: { type: boolean }
        "403":
          { description: The email is not an active administrator invitation }
        "409":
          {
            description: The subject belongs to another user or the invitation is linked to another identity,
          }
...
  /users:
    get:
      summary: List users
      responses:
        "200": { description: Users }
    post:
      summary: Register an unlinked user invitation
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateUser" }
      responses:
        "201": { description: Created }
        "409": { description: A user with the same email already exists }
...
  /service-accounts:
    get:
      summary: List service accounts
      responses:
        "200": { description: Service accounts }
    post:
      summary: Create a service account
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateServiceAccount" }
      responses:
        "201": { description: Created }
  /service-accounts/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      summary: Get a service account
      responses:
        "200": { description: Service account }
    patch:
      summary: Update a service account
      responses:
        "200": { description: Updated }
    delete:
      summary: Disable a service account and revoke all related API keys
      responses:
        "204": { description: Disabled and keys revoked }
...
  /service-accounts/{id}/api-keys:
    post:
      summary: Issue an API key for a service account
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateApiKey" }
      responses:
        "201": { description: Key issued. The secret is returned only once. }
...
  /jobs/render-tickets:
    post:
      summary: Create a render job from the Access-protected administrator Web UI
      parameters:
        - in: header
          name: X-CSRF-Token
          required: true
          schema: { type: string, const: "1" }
        - in: header
          name: Idempotency-Key
          required: true
          schema: { type: string, minLength: 16, maxLength: 200 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminRenderTicketRequest" }
      responses:
        "201":
          description: Job reserved and short-lived tickets issued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RenderTicket" }
        "200":
          description: Existing idempotent job and freshly issued short-lived tickets
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RenderTicket" }
        "409":
          {
            description: The selected render identity is unavailable or the idempotency key conflicts,
          }
  /jobs/source-tickets:
    post:
      summary: Reserve or reuse a Source for administrator browser rendering
      parameters:
        - in: header
          name: X-CSRF-Token
          required: true
          schema: { type: string, const: "1" }
        - in: header
          name: Idempotency-Key
          required: true
          schema: { type: string, minLength: 16, maxLength: 200 }
      responses:
        "200": { description: Existing ready or idempotent Source }
        "201": { description: Source reserved }
...
  /jobs/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      summary: Get a render job
      responses:
        "200": { description: Job }
    delete:
      summary: Mark a terminal render job for deletion
      responses:
        "202": { description: Deletion requested }
  /jobs/{id}/cancel:
    post:
      summary: Request cancellation
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "202": { description: Cancellation requested }
  /jobs/{id}/retry:
    post:
      summary: Retry a terminal job
      parameters:
        - $ref: "#/components/parameters/Id"
        - in: header
          name: Idempotency-Key
          required: true
          schema: { type: string, minLength: 16, maxLength: 200 }
      responses:
        "202": { description: Retry created }
...
  /jobs/{id}/previews/{name}:
    get:
      summary: View or download a retained page preview through the administrator session
      parameters:
        - $ref: "#/components/parameters/Id"
        - in: path
          name: name
          required: true
          schema: { type: string, pattern: "^page-[1-9][0-9]{0,2}\\.png$" }
        - in: query
          name: disposition
          schema: { type: string, enum: [attachment, inline] }
      responses:
        "200": { description: Integrity-labelled PNG stream }
        "410": { description: Job artifact retention has expired }
...
  /sources/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      summary: Get Source paths, entrypoints, retention, and related Jobs
      responses:
        "200": { description: Source }
    delete:
      summary: Delete an unreferenced retained Source
      responses:
        "202": { description: Deletion requested }
        "409": { description: Retained Jobs still reference the Source }
  /sources/{id}/render:
    post:
      summary: Create a new Job from an existing Source and selected entrypoint
      parameters:
        - $ref: "#/components/parameters/Id"
        - in: header
          name: Idempotency-Key
          required: true
          schema: { type: string, minLength: 16, maxLength: 200 }
      responses:
        "201": { description: Job created }
  /sources/{id}/source-ref:
    post:
      summary: Issue a short-lived Remote MCP reference for a retained Source
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "201": { description: Source reference issued }
...
  /audit-logs:
    get:
      summary: Search and page audit log entries
      parameters:
        - {
            in: query,
            name: page,
            schema: { type: integer, minimum: 1, default: 1 },
          }
        - {
            in: query,
            name: pageSize,
            schema: { type: integer, minimum: 1, maximum: 100, default: 50 },
          }
        - { in: query, name: action, schema: { type: string, maxLength: 200 } }
        - { in: query, name: actor, schema: { type: string, maxLength: 500 } }
        - { in: query, name: target, schema: { type: string, maxLength: 500 } }
        - { in: query, name: result, schema: { type: string, maxLength: 100 } }
        - { in: query, name: from, schema: { type: string, format: date-time } }
        - { in: query, name: to, schema: { type: string, format: date-time } }
      responses:
        "200":
          {
            description: Audit log page with items,
            page,
            pageSize,
            total,
            and totalPages,
          }
components:
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema: { type: string }
  schemas:
    Session:
      type: object
      required: [identity, linkage]
      properties:
        identity:
          type: object
          required: [subject, email]
          properties:
            subject: { type: string }
            email: { type: [string, "null"], format: email }
        linkage:
          type: object
          required: [state, claimable, user]
          properties:
            state:
              type: string
              enum:
                [
                  identity_incomplete,
                  not_invited,
                  disabled,
                  ineligible,
                  claimable,
                  linked,
                  conflict,
                ]
            claimable: { type: boolean }
            user:
              oneOf:
                - type: "null"
                - type: object
                  required:
                    [id, email, displayName, role, status, linkedAt, generation]
                  properties:
                    id: { type: string }
                    email: { type: string, format: email }
                    displayName: { type: string }
                    role: { $ref: "#/components/schemas/Role" }
                    status: { type: string, enum: [active, disabled] }
                    linkedAt: { type: [string, "null"], format: date-time }
                    generation: { type: integer, minimum: 0 }
    Role:
      type: string
      enum: [user, admin, owner]
    CreateUser:
      type: object
      additionalProperties: false
      required: [email, displayName, role]
      properties:
        email: { type: string, format: email, maxLength: 320 }
        displayName: { type: string, minLength: 1, maxLength: 200 }
        role: { $ref: "#/components/schemas/Role" }
    CreateServiceAccount:
      type: object
      additionalProperties: false
      required: [ownerUserId, name, clientType]
      properties:
        ownerUserId: { type: string }
        name: { type: string, minLength: 1, maxLength: 100 }
        clientType:
          { type: string, enum: [codex, claude-code, mcp, ci, generic] }
    CreateApiKey:
      type: object
      additionalProperties: false
      required: [name, scopes]
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        scopes:
          type: array
          minItems: 1
          maxItems: 32
          uniqueItems: true
          description: Render and admin scopes cannot be mixed. Admin key issuance is owner-only.
          items:
            type: string
            enum:
              [
                render:create,
                render:read:own,
                "admin:users:read",
                "admin:users:write",
                "admin:service-accounts:read",
                "admin:service-accounts:write",
                "admin:api-keys:read",
                "admin:api-keys:write",
                "admin:jobs:read",
                "admin:jobs:write",
                "admin:system:read",
                "admin:system:write",
                "admin:audit:read",
                "admin:*",
              ]
        expiresAt: { type: [string, "null"], format: date-time }
...
    AdminRenderTicketRequest:
      oneOf:
        - type: object
          additionalProperties: false
          required: [apiKeyId, size, sha256]
          properties:
            apiKeyId: { type: string, minLength: 1, maxLength: 128 }
            size: { type: integer, minimum: 1, maximum: 20971520 }
            sha256: { type: string, pattern: "^[a-f0-9]{64}$" }
            outputs: { $ref: "#/components/schemas/RenderOutputs" }
        - type: object
          additionalProperties: false
          required: [apiKeyId, sourceId]
          properties:
            apiKeyId: { type: string, minLength: 1, maxLength: 128 }
            sourceId: { type: string, pattern: "^source_[a-f0-9]{32}$" }
            entrypoint: { type: string, maxLength: 2048, default: main.tex }
            outputs: { $ref: "#/components/schemas/RenderOutputs" }
    RenderOutputs:
      type: array
      default: [pdf]
      minItems: 1
      maxItems: 2
      uniqueItems: true
      items: { type: string, enum: [pdf, svg] }
    RenderTicket:
      type: object
      additionalProperties: false
      required: [jobId, uploadTicket, jobTicket, uploadUrl, expiresAt]
      properties:
        jobId: { type: string, pattern: "^job_[a-f0-9]{32}$" }
        uploadTicket: { type: string }
        jobTicket: { type: string }
        uploadUrl: { type: string, format: uri }
        expiresAt: { type: string, format: date-time }
  securitySchemes:
    cloudflareAccess:
      type: apiKey
      in: header
      name: Cf-Access-Jwt-Assertion
      description: Cloudflare Access JWT verified by the server. Its sub and email claims are authoritative.
    accessAndAdmin:
      type: http
      scheme: bearer
      description: Cloudflare Access is always required. The administration CLI also supplies an lra_ API key.